requirements.txt
albumentations==0.4.3
biosppy
efficientnet_pytorch==0.6.3
future
matplotlib==3.0.3
numpy==1.21.0
opencv_python==4.2.0.32
pandas==0.25.3
Pillow<7
plotly==4.7.0
PyQt5
PyYAML
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorboard==1.14
torch==1.1.0
torchvision==0.3.0
tqdm==4.42.1
wfdb==2.2.1
